backend/requirements.txt
requests==2.28.1
pytest==7.4.0
allure-pytest==2.10.0
python-dotenv==1.0.0